Refactor shape_inference_test.py  (#4302)
* Split shape_inference_test.py

Move testing shape inference with data propagation in a separate
file 'data_propagation_test.py'.


* Move common functions to TestShapeInferenceHelper class

Create a helper class with the common functionality between
TestShapeInference and TestDataPropagation and let both classes
inherit it.


* Remove unused imports


* Use full import paths in data_propagation_test


* Revert to local import

Due to CI failures when using full path imports from another
module in the test direcroty, revert to local import.

Ignore the mypy check since the current mypy version cannot
find the imported module, potentially newer versions can solve the issue
by using the '--namespace-packages' option.
